Use lib.es5.d.ts instead of lib.d.ts.  Delete collections.d.ts
There's a way to tell TypeScript 1.5.0-beta to recognize ES6 but still
target ES5.  To do this, we pass the `--noLib` flag to the compiler to
not use the default definitions (which would be `lib.d.ts`) and provide
`lib.es6.d.ts` as an addition and first input file.

This is cleaner than maintaining collections.d.ts and have less trouble
with newer TypeScript versions.